+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Add pagination query parametrs into /admin/merchants route
   - add search query parametr to make search by Store name or Email
 
+### Bug fixes
+
+- Bug fixes list products `productStoreAccessMiddleware`
+
 ## 0.38.0
 
 ### Bug fixes